pyproject.toml
[project]
name = "converse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.13.4",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20260508",
]
